Fix v3 build errors and replace zeroshot phases with v4 plan (3 phases)
- Fix missing WaiveFonts.h include in WaiveLookAndFeel.cpp
- Fix ColourPalette::error → danger in ConsoleComponent.cpp
- Fix Fonts::small() → Fonts::meter() in MixerChannelStrip.cpp
- Fix fillTriangle → Path+fillPath in ClipComponent.cpp
- Fix ScrollBar::setTooltip() removal in TimelineComponent.cpp
- Fix deprecated Font constructor in WaiveFonts.h and ToolSidebarComponent.cpp
- Fix sign-conversion warnings in MixerComponent.cpp
- Remove PCH (conflicts with JUCE module .cpp compilation)
- New v4 phases: hardcoded colors, performance, UI polish

+++ gui/src/ui/MixerComponent.cpp
@@ -161,8 +161,8 @@ bool MixerComponent::keyPressed (const juce::KeyPress& key)
         if (focusedStripIndex > 0)
         {
             focusedStripIndex--;
-            if (focusedStripIndex < (int) strips.size() && strips[focusedStripIndex] != nullptr)
-                strips[focusedStripIndex]->grabKeyboardFocus();
+            if (focusedStripIndex < (int) strips.size() && strips[(size_t) focusedStripIndex] != nullptr)
+                strips[(size_t) focusedStripIndex]->grabKeyboardFocus();
             return true;
         }
     }
@@ -171,8 +171,8 @@ bool MixerComponent::keyPressed (const juce::KeyPress& key)
         if (focusedStripIndex < (int) strips.size() - 1)
         {
             focusedStripIndex++;
-            if (strips[focusedStripIndex] != nullptr)
-                strips[focusedStripIndex]->grabKeyboardFocus();
+            if (strips[(size_t) focusedStripIndex] != nullptr)
+                strips[(size_t) focusedStripIndex]->grabKeyboardFocus();
             return true;
         }
         else if (focusedStripIndex == (int) strips.size() - 1 && masterStrip != nullptr)
